All basic CRUD operations implemented for Listing

+++ findmyparadise/src/main/java/com/tp/wrc/findmyparadise/services/ListingServiceImpl.java
@@ -35,7 +35,10 @@ public class ListingServiceImpl implements ListingService {
     }
 
     @Override
-    public boolean destroy(Integer id) {
+    public boolean destroy(Integer id) throws NullListingIDException {
+        if (id == null) {
+            throw new NullListingIDException("Listing ID cannot be null");
+        }
         Listing listing = repo.findById(id).get();
         if (listing != null) {
             repo.delete(listing);
@@ -45,8 +48,32 @@ public class ListingServiceImpl implements ListingService {
         }
     }
 
+    @Override
+    public Listing update(Integer id, Listing newListing) throws NullListingIDException, NoListingFoundException {
+        if (id == null) {
+            throw new NullListingIDException("Listing ID cannot be null");
+        }
 
-//    TODO: implement these methods
+        Listing toUpdate = show(id);
+
+        toUpdate.setName(newListing.getName());
+        toUpdate.setLatitude(newListing.getLatitude());
+        toUpdate.setLongitude(newListing.getLongitude());
+        toUpdate.setHost(newListing.getHost());
+        toUpdate.setAddress(newListing.getAddress());
+        toUpdate.setDescription(newListing.getDescription());
+        toUpdate.setPrice(newListing.getPrice());
+        toUpdate.setMaxGuests(newListing.getMaxGuests());
+        toUpdate.setServiceFee(newListing.getServiceFee());
+        toUpdate.setOccupancyFee(newListing.getOccupancyFee());
+        toUpdate.setCleaningFee(newListing.getCleaningFee());
+        toUpdate.setReviews(newListing.getReviews());
+        toUpdate.setReservations(newListing.getReservations());
+
+        return toUpdate;
+    }
+
+    //    TODO: implement these methods
       @Override
       public Listing findByNameIgnoreCase(String listingName) throws NoListingFoundException, InvalidListingNameException {
           return null;
